calculateLegend
