+++ packages/desktop-client/src/components/reports/spreadsheets/calculateLegend.ts
@@ -1,0 +1,31 @@
+import { theme } from '../../../style';
+import { getColorScale } from '../chart-theme';
+import { type ItemEntity, type MonthData } from '../entities';
+
+function calculateLegend(
+  monthData: MonthData[],
+  calcDataFiltered: ItemEntity[],
+  groupBy: string,
+  graphType: string,
+  balanceTypeOp: string,
+) {
+  const colorScale = getColorScale('qualitative');
+  const chooseData = ['Month', 'Year'].includes(groupBy)
+    ? monthData
+    : calcDataFiltered;
+  return chooseData.map((c, index) => {
+    return {
+      name: ['Month', 'Year'].includes(groupBy) ? c.date : c.name,
+      color:
+        graphType === 'DonutGraph'
+          ? colorScale[index % colorScale.length]
+          : ['Month', 'Year'].includes(groupBy)
+          ? balanceTypeOp === 'totalDebts'
+            ? theme.reportsRed
+            : theme.reportsBlue
+          : colorScale[index % colorScale.length],
+    };
+  });
+}
+
+export default calculateLegend;
